% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multip.R
\name{multip}
\alias{multip}
\title{Function to multiply two values}
\usage{
multip(x, y)
}
\arguments{
\item{x}{this is the first value we want to multiply.}

\item{y}{this is the second value we want to multiply.}
}
\description{
This function multiplies two values
}
\examples{
# Primer ejemplo
multip(5, 10)
# Segundo ejemplos
a <- 5
b <- 15
res <- multip(x=a, y=b)
res

}
