% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Function to add two values}
\usage{
suma(x, y)
}
\arguments{
\item{x}{this is the first value we want to add.}

\item{y}{this is the second value we want to add.}
}
\description{
This function adds two values
}
\examples{
# Primer ejemplo
suma(5, 10)
# Segundo ejemplos
a <- 5
b <- 15
res <- suma(x=a, y=b)
res

}
